# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/users/search": {
      "get": {
        "operationId": "UsersController_searchUsers_v1",
        "parameters": [
          {
            "name": "email",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Search users by email",
        "tags": [
          "users"
        ]
      }
    },
    "/v1/users/profile": {
      "get": {
        "operationId": "UsersController_getProfile_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Profile retrieved."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_updateProfile_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update user profile",
        "tags": [
          "users"
        ]
      }
    },
    "/v1/users/change-password": {
      "post": {
        "operationId": "UsersController_changePassword_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change password",
        "tags": [
          "users"
        ]
      }
    },
    "/v1/users/organizations": {
      "get": {
        "operationId": "UsersController_getMyOrganizations_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Organizations retrieved."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user organizations",
        "tags": [
          "users"
        ]
      }
    },
    "/v1/users/switch-context": {
      "post": {
        "operationId": "UsersController_switchContext_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchContextDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Context switched."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Switch active organization context",
        "tags": [
          "users"
        ]
      }
    },
    "/v1/organizations": {
      "post": {
        "operationId": "OrganizationsController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Organization created."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create organization",
        "tags": [
          "organizations"
        ]
      },
      "get": {
        "operationId": "OrganizationsController_findAll_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Organizations retrieved."
          }
        },
        "summary": "Get all organizations",
        "tags": [
          "organizations"
        ]
      }
    },
...
    "/v1/payments/{id}": {
      "get": {
        "operationId": "PaymentsController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Payments"
        ]
      },
      "patch": {
        "operationId": "PaymentsController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Payments"
        ]
      },
      "delete": {
        "operationId": "PaymentsController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Payments"
        ]
      }
    },
...
  },
  "info": {
    "title": "Shadboard API",
    "description": "The Shadboard API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "imageProfileUrl": {
            "type": "string"
          }
        }
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "SwitchContextDto": {
        "type": "object",
        "properties": {
          "organizationId": {
            "type": "string"
          }
        },
        "required": [
          "organizationId"
        ]
      },
      "CreateOrganizationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
...
    }
  }
}